Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
bcrypt>=4.0.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
email-validator>=2.0.0
